feat: add movies app JavaScript code
Added code for adding event listeners on and handlers for the submit
and remove buttons.

Added code to add and remove movies from the movies database (Map).

Added code for creating and displaying the movies in the HTML.

In the HTML, the unordered list HTML element for listing movies was
changed to a table, to more easily display and organize the names and
ratings.

A Map was used to store movie info, because it prevents duplicates.  If
no data structure was used, then the whole movies table would need to
be iterated thru for every addition of a movie to check for existence.

In displayMovies(), the line of code to empty out the movies table is
for a future commit involving sorting.