feat: add sorting
* Added event listeners for clicking on the table headers and handlers
for processing the sorting choice.

* Added methods for changing the sorting type and for doing the sort.

The sorting is applied when clicking on the table headers, because that
seems to be the more intuitive action, when compared to adding buttons.

The sorting is done for each movie submission, so that the list of
movies is always sorted.  This seems to be the more common expected
behavior.

The algorithm for sortMovies() was taken and adapted from a
geeksforgeeks article.  There is some repeated code when assigning the
comparator variable, but the only alternative I can think of is to put
the switch statement within the comparator function.  For that case
though, it would cause the switch statement to check for the sort type
for every movie comparison, which adds extra processing when compared
to just doing the same type of compare.  Since memory space is cheap in
comparison to processing time, the currently implemented method was the
best approach.  sortMovies() can be refactored, if there is a way to
reduce repeated code without changing functionality, or in other words,
code flow.